llm_vision_models/requirements.txt
fastapi
uvicorn
Pillow
torch
transformers
accelerate
huggingface_hub
requests
pydantic
colpali-engine
stamina
tqdm
ollama
requests